pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
